username validation

+++ routes/ideas.js
@@ -50,17 +50,26 @@ router.post('/', async (request, response) => {
 // Update idea
 router.put('/:id', async (request, response) => {
   try {
-    const updatedIdea = await Idea.findByIdAndUpdate(
-      request.params.id,
-      {
-        $set: {
-          text: request.body.text,
-          tag: request.body.tag,
+    const idea = await Idea.findById(request.params.id);
+    //check usernames
+    if (idea.username === request.body.username) {
+      const updatedIdea = await Idea.findByIdAndUpdate(
+        request.params.id,
+        {
+          $set: {
+            text: request.body.text,
+            tag: request.body.tag,
+          },
         },
-      },
-      { new: true }
-    );
-    response.json({ success: true, data: updatedIdea });
+        { new: true }
+      );
+      return response.json({ success: true, data: updatedIdea });
+    }
+    // usernames do not match
+    response.status(403).json({
+      success: false,
+      error: 'You are not authorised to update this resource',
+    });
   } catch (error) {
     console.log(error);
     response
@@ -72,8 +81,19 @@ router.put('/:id', async (request, response) => {
 // Delete idea
 router.delete('/:id', async (request, response) => {
   try {
-    await Idea.findByIdAndDelete(request.params.id);
-    response.json({ success: true, data: {} });
+    const idea = await Idea.findById(request.params.id);
+
+    // Match the usernames
+    if (idea.username === request.body.username) {
+      await Idea.findByIdAndDelete(request.params.id);
+      return response.json({ success: true, data: {} });
+    }
+
+    //usernames do not match
+    response.status(403).json({
+      success: false,
+      error: 'You are not authorised to delete this resource',
+    });
   } catch (error) {
     console.log(error);
     response
